refactor($baseball) 테스트 위치 정리
