+++ src/test/java/baseball/model/NumberToListChangerTest.java
@@ -1,8 +1,7 @@
-package baseball.baseballgame;
+package baseball.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import baseball.model.NumberToListChanger;
 import java.util.List;
 import org.junit.jupiter.api.Test;
