+++ src/test/java/baseball/model/ball/BallTest.java
@@ -1,9 +1,7 @@
-package baseball.baseballgame.ball;
+package baseball.model.ball;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import baseball.model.ball.Ball;
-import baseball.model.ball.BallNumber;
 import org.junit.jupiter.api.Test;
 
 public class BallTest {
